TextInputSectionを共通Widet化
