+++ packages/mottai_flutter_app/lib/development/review/ui/review_form.dart
@@ -5,7 +5,7 @@ import 'package:gap/gap.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../../review/ui/review_controller.dart';
-import '../../../widgets/optional_badge.dart';
+import '../../../widgets/text_input_section.dart';
 import '../../firebase_storage/firebase_storage.dart';
 import '../../firebase_storage/ui/firebase_storage_controller.dart';
 
@@ -107,7 +107,7 @@ class ReviewFormState extends ConsumerState<ReviewForm> {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  _TextInputSection(
+                  TextInputSection(
                     title: 'タイトル',
                     description: '感想のタイトルを最大2行程度で入力してください。',
                     maxLines: 2,
@@ -115,7 +115,7 @@ class ReviewFormState extends ConsumerState<ReviewForm> {
                     controller: _titleController,
                     isRequired: true,
                   ),
-                  _TextInputSection(
+                  TextInputSection(
                     title: '本文',
                     description: '感想の本文を入力してください。',
                     defaultDisplayLines: 5,
@@ -162,108 +162,6 @@ class ReviewFormState extends ConsumerState<ReviewForm> {
               ),
             ),
           ),
-        ],
-      ),
-    );
-  }
-}
-
-//TODO JobForm で使用しているものと同様のため、dart_flutter_common で共通化すべきか？
-/// タイトルと説明、テキストフィールドからなるセクション。
-/// [Section] を使用し、`Section.content` にフィールドを与えている。
-class _TextInputSection<T extends dynamic> extends StatelessWidget {
-  /// テキスト入力のみをさせる通常の [_TextInputSection] を作成する。
-  const _TextInputSection({
-    required this.title,
-    this.description,
-    this.maxLines,
-    this.defaultDisplayLines = 1,
-    this.controller,
-    this.isRequired = false,
-  })  : choices = const {},
-        enabledChoices = const [],
-        onChoiceSelected = null;
-
-  /// テキスト入力と選択肢を併せて入力させる [_TextInputSection] を作成する。
-  const _TextInputSection.withChoice({
-    required this.title,
-    this.description,
-    this.maxLines,
-    this.defaultDisplayLines = 1,
-    required this.choices,
-    required this.enabledChoices,
-    required this.onChoiceSelected,
-    this.controller,
-    this.isRequired = false,
-  });
-
-  /// セクションのタイトル。
-  final String title;
-
-  /// セクションの説明。
-  final String? description;
-
-  /// テキストフィールドの最大行数
-  final int? maxLines;
-
-  /// 初期表示時のテキストフィールドの行数
-  final int defaultDisplayLines;
-
-  /// テキストフィールドの下に表示する選択肢
-  /// 選択された際の値がkeyで、表示する値がvalueの[Map]で受け取る。
-  final Map<T, String> choices;
-
-  /// [choices] の有効な値のリスト
-  final List<T> enabledChoices;
-
-  /// [choices] が選択された際のコールバック
-  final void Function(T item)? onChoiceSelected;
-
-  /// テキストフィールドのコントローラー
-  final TextEditingController? controller;
-
-  /// 必須入力か否か
-  final bool isRequired;
-
-  @override
-  Widget build(BuildContext context) {
-    return Section(
-      title: title,
-      titleBadge: Padding(
-        padding: const EdgeInsets.only(left: 8),
-        child: OptionalBadge(isRequired: isRequired),
-      ),
-      titleStyle: Theme.of(context).textTheme.titleLarge,
-      description: description,
-      descriptionStyle: Theme.of(context).textTheme.bodyMedium,
-      sectionPadding: const EdgeInsets.only(bottom: 32),
-      content: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          TextFormField(
-            controller: controller,
-            maxLines: maxLines,
-            validator: (value) {
-              if (isRequired && (value ?? '').isEmpty) {
-                return '入力してください。';
-              }
-              return null;
-            },
-            decoration: InputDecoration(
-              hintText: List.filled(defaultDisplayLines - 1, '\n').join(),
-              border: const OutlineInputBorder(),
-            ),
-          ),
-          if (choices.isNotEmpty) ...[
-            const Gap(16),
-            SelectableChips<T>(
-              allItems: choices.keys,
-              labels: choices,
-              runSpacing: 8,
-              enabledItems: enabledChoices,
-              onTap: onChoiceSelected,
-            ),
-          ],
         ],
       ),
     );
